Move PlayerInfo class from game.ts to new file player.ts
